fix load balancer: singleflight dedup, scored candidates, Int64 schema types
Fixes from gemini-code-assist review:
- Fix dbQueried flag tracking with Is/Mark methods for RPM and TPM
- Add singleflight dedup for concurrent quota DB queries
- Eliminate double Score() calls via SortWithScores and scored candidates
- Fix GraphQL Int overflow by switching to Int64 for rpm/tpm types
- Track HardExhausted flag on scored candidates for proper filtering
- Fix singleflight cache keys to include window anchor timestamps
- Restore input validation by moving helper functions to resolver_helpers

Addresses: dbQueried dedup, singleflight cache keys, double Score elimination,
Int64 schema overflow, HardExhausted flag, anchor-in-singleflight-keys,
restored input validation

+++ internal/server/gql/models_gen.go
@@ -121,12 +121,12 @@ type ChannelPerformanceStat struct {
 }
 
 type ChannelRateLimitStatus struct {
-	RpmCurrent        *int       `json:"rpmCurrent,omitempty"`
-	RpmLimit          *int       `json:"rpmLimit,omitempty"`
+	RpmCurrent        *int64     `json:"rpmCurrent,omitempty"`
+	RpmLimit          *int64     `json:"rpmLimit,omitempty"`
 	RpmResetAt        *time.Time `json:"rpmResetAt,omitempty"`
 	RpmWindowAnchor   *time.Time `json:"rpmWindowAnchor,omitempty"`
-	TpmCurrent        *int       `json:"tpmCurrent,omitempty"`
-	TpmLimit          *int       `json:"tpmLimit,omitempty"`
+	TpmCurrent        *int64     `json:"tpmCurrent,omitempty"`
+	TpmLimit          *int64     `json:"tpmLimit,omitempty"`
 	TpmResetAt        *time.Time `json:"tpmResetAt,omitempty"`
 	TpmWindowAnchor   *time.Time `json:"tpmWindowAnchor,omitempty"`
 	CostCurrent       *float64   `json:"costCurrent,omitempty"`
